[vim] move lsp conf file to plugs dir
